sentiment analysis improved
