--- lib/agents/social-agent.ts
@@ -1,750 +1,0 @@
-// NOTE: This agent now uses only CdpV2EvmWalletProvider for wallet management (Coinbase AgentKit v2)
-import { DecodedMessage } from '@xmtp/browser-sdk';
-import { DynamicStructuredTool } from '@langchain/core/tools';
-import { z } from 'zod';
-import { BaseAgent } from './base-agent';
-import { 
-  AgentKit,
-  CdpV2EvmWalletProvider,
-  walletActionProvider,
-  cdpApiActionProvider,
-  messariActionProvider
-} from '@coinbase/agentkit';
-import { getLangChainTools } from '@coinbase/agentkit-langchain';
-import { HumanMessage } from '@langchain/core/messages';
-import { MemorySaver } from '@langchain/langgraph';
-import { createReactAgent } from '@langchain/langgraph/prebuilt';
-import { ChatOpenAI } from '@langchain/openai'
-import axios from 'axios';
-import {
-  SocialAgentConfig,
-  AgentContext,
-  AgentResponse,
-  CuratedContent,
-  ContentSource,
-  UserPreferences,
-  EngagementMetrics,
-} from '../types';
-
-/**
- * Production-grade SocialAgent with real content curation and blockchain tipping
- * Uses createReactAgent for proper LLM + tool integration
- */
-export class SocialAgent extends BaseAgent {
-  private agentKit?: AgentKit;
-  private walletProvider?: CdpV2EvmWalletProvider;
-  private reactAgent?: ReturnType<typeof createReactAgent>;
-  private memory?: MemorySaver;
-  private llmModel?: ChatOpenAI;
-  private contentSources: Map<string, ContentSource> = new Map();
-  private curatedContent: Map<string, CuratedContent[]> = new Map();
-  private userPreferences: Map<string, UserPreferences> = new Map();
-
-  constructor(config: SocialAgentConfig) {
-    super(config);
-    this.initializeContentSources();
-  }
-
-  /**
-   * Initialize AgentKit with createReactAgent for proper LLM integration
-   */
-  async initialize(): Promise<void> {
-    try {
-      // Initialize LLM
-      this.llmModel = new ChatOpenAI({
-        model: "gpt-4o",
-        temperature: 0.3,
-      });
-
-      // Configure CDP Wallet Provider for social tipping
-      const cdpWalletConfig = {
-        apiKeyId: process.env.CDP_API_KEY_ID!,
-        apiKeySecret: process.env.CDP_API_KEY_SECRET!,
-        walletSecret: process.env.CDP_WALLET_SECRET!,
-        idempotencyKey: process.env.IDEMPOTENCY_KEY,
-        address: process.env.ADDRESS as `0x${string}` | undefined,
-        networkId: process.env.NETWORK_ID!,
-      };
-
-      this.walletProvider = await CdpV2EvmWalletProvider.configureWithWallet(cdpWalletConfig);
-
-      // Initialize AgentKit for social features
-      this.agentKit = await AgentKit.from({
-        walletProvider: this.walletProvider,
-        actionProviders: [
-          walletActionProvider(),
-          cdpApiActionProvider({
-            apiKeyId: process.env.CDP_API_KEY_ID!,
-            apiKeySecret: process.env.CDP_API_KEY_PRIVATE_KEY!,
-          }),
-          messariActionProvider({
-            apiKey: process.env.MESSARI_API_KEY!,
-          }),
-        ],
-      });
-
-      // Get AgentKit tools for LangChain integration
-      const agentKitTools = await getLangChainTools(this.agentKit);
-      
-      // Add custom tools
-      this.initializeTools();
-      const allTools = [...agentKitTools, ...this.tools];
-
-      // Initialize memory for conversation context
-      this.memory = new MemorySaver();
-
-      // Create React Agent with LLM, tools, and memory
-      this.reactAgent = createReactAgent({
-        llm: this.llmModel,
-        tools: allTools,
-        checkpointSaver: this.memory,
-        messageModifier: this.getSystemPrompt(),
-      });
-
-      await super.initialize();
-      this.logger.info('SocialAgent initialized with createReactAgent and real content capabilities');
-    } catch (error) {
-      this.logger.error('Failed to initialize SocialAgent with AgentKit', { error });
-      throw error;
-    }
-  }
-
-  protected initializeTools(): void {
-    this.tools.push(
-      new DynamicStructuredTool({
-        name: 'get_real_crypto_news',
-        description: 'Get real-time cryptocurrency news from multiple sources',
-        schema: z.object({
-          category: z.string().optional(),
-          limit: z.number().optional().default(5),
-        }),
-        func: async ({ category, limit }) => {
-          try {
-            const news = await this.getRealCryptoNews(category, limit);
-            return `📰 Latest Crypto News:\n${news.map(n => 
-              `• ${n.title}\n  Source: ${n.source}\n  ${n.content.substring(0, 100)}...`
-            ).join('\n\n')}`;
-          } catch (error) {
-            this.logger.error('Error fetching crypto news', { error });
-            return `Error fetching crypto news: ${error instanceof Error ? error.message : 'Unknown error'}`;
-          }
-        },
-      }),
-
-      new DynamicStructuredTool({
-        name: 'get_real_trending_topics',
-        description: 'Get real trending topics from social media and crypto platforms',
-        schema: z.object({
-          platform: z.enum(['twitter', 'reddit', 'coingecko', 'all']).optional().default('all'),
-        }),
-        func: async ({ platform }) => {
-          try {
-            const trends = await this.getRealTrendingTopics(platform);
-            return `🔥 Trending Now:\n${trends.map((trend, i) => `${i+1}. ${trend}`).join('\n')}`;
-          } catch (error) {
-            this.logger.error('Error fetching trending topics', { error });
-            return `Error fetching trending topics: ${error instanceof Error ? error.message : 'Unknown error'}`;
-          }
-        },
-      }),
-
-      new DynamicStructuredTool({
-        name: 'analyze_real_sentiment',
-        description: 'Analyze real sentiment using AI-powered sentiment analysis',
-        schema: z.object({
-          text: z.string(),
-          context: z.string().optional(),
-        }),
-        func: async ({ text, context }) => {
-          try {
-            const sentiment = await this.analyzeRealSentiment(text, context);
-            return `📊 Sentiment Analysis:
-Sentiment: ${sentiment.overall} (${sentiment.confidence}% confidence)
-Emotions: ${sentiment.emotions.join(', ')}
-Key Phrases: ${sentiment.keyPhrases.join(', ')}`;
-          } catch (error) {
-            this.logger.error('Error analyzing sentiment', { error });
-            return `Error analyzing sentiment: ${error instanceof Error ? error.message : 'Unknown error'}`;
-          }
-        },
-      }),
-
-      new DynamicStructuredTool({
-        name: 'tip_content_creator',
-        description: 'Tip a content creator with USDC for valuable content',
-        schema: z.object({
-          creatorAddress: z.string(),
-          amount: z.number(),
-          contentId: z.string().optional(),
-          message: z.string().optional(),
-        }),
-        func: async ({ creatorAddress, amount, contentId, message }) => {
-          try {
-            const txHash = await this.tipCreator(creatorAddress, amount, contentId, message);
-            return `💰 Tip sent successfully!
-Amount: ${amount} USDC
-To: ${creatorAddress}
-Transaction: ${txHash}
-${message ? `Message: ${message}` : ''}`;
-          } catch (error) {
-            this.logger.error('Error sending tip', { error });
-            return `Error sending tip: ${error instanceof Error ? error.message : 'Unknown error'}`;
-          }
-        },
-      }),
-
-      new DynamicStructuredTool({
-        name: 'get_market_sentiment',
-        description: 'Get real market sentiment for specific cryptocurrencies',
-        schema: z.object({
-          tokens: z.array(z.string()).optional().default(['bitcoin', 'ethereum']),
-        }),
-        func: async ({ tokens }) => {
-          try {
-            const sentiments = await Promise.all(
-              tokens.map(async (token: string) => {
-                const sentiment = await this.getTokenSentiment(token);
-                return `${token.toUpperCase()}: ${sentiment.score}% positive (${sentiment.source})`;
-              })
-            );
-            return `📈 Market Sentiment:\n${sentiments.join('\n')}`;
-          } catch (error) {
-            this.logger.error('Error fetching market sentiment', { error });
-            return `Error fetching market sentiment: ${error instanceof Error ? error.message : 'Unknown error'}`;
-          }
-        },
-      }),
-
-      new DynamicStructuredTool({
-        name: 'create_content_summary',
-        description: 'Create AI-powered summaries of crypto content',
-        schema: z.object({
-          urls: z.array(z.string()),
-          topic: z.string().optional(),
-        }),
-        func: async ({ urls, topic }) => {
-          try {
-            const summaries = await Promise.all(
-              urls.map(async (url: string) => {
-                const summary = await this.createContentSummary(url, topic);
-                return `📄 ${summary.title}\n${summary.summary}`;
-              })
-            );
-            return `📝 Content Summaries:\n\n${summaries.join('\n\n')}`;
-          } catch (error) {
-            this.logger.error('Error creating content summaries', { error });
-            return `Error creating content summaries: ${error instanceof Error ? error.message : 'Unknown error'}`;
-          }
-        },
-      }),
-
-      new DynamicStructuredTool({
-        name: 'track_social_metrics',
-        description: 'Track social engagement metrics and viral content',
-        schema: z.object({
-          contentType: z.enum(['news', 'defi', 'nft', 'meme']).optional().default('news'),
-          timeframe: z.enum(['1h', '24h', '7d']).optional().default('24h'),
-        }),
-        func: async ({ contentType, timeframe }) => {
-          try {
-            const metrics = await this.trackSocialMetrics(contentType, timeframe);
-            return `📊 Social Metrics (${timeframe}):
-Top Content: ${metrics.topContent.title}
-Engagement: ${metrics.totalEngagement} interactions
-Viral Score: ${metrics.viralScore}/100
-Trending Hashtags: ${metrics.trendingHashtags.join(', ')}`;
-          } catch (error) {
-            this.logger.error('Error tracking social metrics', { error });
-            return `Error tracking social metrics: ${error instanceof Error ? error.message : 'Unknown error'}`;
-          }
-        },
-      }),
-    );
-  }
-
-  /**
-   * Get real cryptocurrency news from multiple sources
-   */
-  private async getRealCryptoNews(category?: string, limit: number = 5): Promise<CuratedContent[]> {
-    try {
-      const news: CuratedContent[] = [];
-
-      // CoinGecko News API
-      try {
-        const response = await axios.get('https://api.coingecko.com/api/v3/news', {
-          params: { per_page: Math.min(limit, 10) }
-        });
-        
-        response.data.data.forEach((item: any) => {
-          news.push({
-            id: item.id,
-            title: item.title,
-            content: item.description || item.title,
-            source: 'CoinGecko',
-            category: category || 'general',
-            relevanceScore: 0.8,
-            timestamp: new Date(item.published_at),
-            url: item.url,
-          });
-        });
-      } catch (error) {
-        this.logger.warn('CoinGecko news API failed', { error });
-      }
-
-      // CryptoCompare News API
-      try {
-        const response = await axios.get('https://min-api.cryptocompare.com/data/v2/news/', {
-          params: { 
-            lang: 'EN',
-            feeds: 'CoinDesk,CoinTelegraph,TheBlock',
-            lmt: Math.min(limit, 10)
-          }
-        });
-
-        response.data.Data.forEach((item: any) => {
-          news.push({
-            id: item.id,
-            title: item.title,
-            content: item.body,
-            source: item.source_info.name,
-            category: category || 'general',
-            relevanceScore: 0.7,
-            timestamp: new Date(item.published_on * 1000),
-            url: item.url,
-          });
-        });
-      } catch (error) {
-        this.logger.warn('CryptoCompare news API failed', { error });
-      }
-
-      // Sort by timestamp and limit
-      return news
-        .sort((a, b) => b.timestamp.getTime() - a.timestamp.getTime())
-        .slice(0, limit);
-    } catch (error) {
-      this.logger.error('Error fetching real crypto news', { error });
-      throw error;
-    }
-  }
-
-  /**
-   * Get real trending topics from multiple platforms
-   */
-  private async getRealTrendingTopics(platform: string = 'all'): Promise<string[]> {
-    try {
-      const trends: string[] = [];
-
-      // CoinGecko trending coins
-      try {
-        const response = await axios.get('https://api.coingecko.com/api/v3/search/trending');
-        const trendingCoins = response.data.coins.slice(0, 5).map((coin: any) => 
-          `$${coin.item.symbol.toUpperCase()}`
-        );
-        trends.push(...trendingCoins);
-      } catch (error) {
-        this.logger.warn('CoinGecko trending failed', { error });
-      }
-
-      // Add general crypto trends
-      const cryptoTrends = [
-        'DeFi Summer 2.0',
-        'Layer 2 Scaling',
-        'NFT Gaming',
-        'RWA Tokenization',
-        'AI + Crypto'
-      ];
-      trends.push(...cryptoTrends);
-
-      return trends.slice(0, 10);
-    } catch (error) {
-      this.logger.error('Error fetching trending topics', { error });
-      throw error;
-    }
-  }
-
-  /**
-   * Real sentiment analysis using AI
-   */
-  private async analyzeRealSentiment(text: string, context?: string): Promise<{
-    overall: string;
-    confidence: number;
-    emotions: string[];
-    keyPhrases: string[];
-  }> {
-    try {
-      // Use LLM for sentiment analysis
-      const prompt = `Analyze the sentiment of this text${context ? ` in the context of ${context}` : ''}:
-"${text}"
-
-Return analysis as JSON with:
-- overall: "positive", "negative", or "neutral"
-- confidence: number 0-100
-- emotions: array of detected emotions
-- keyPhrases: array of important phrases`;
-
-      const response = await this.processWithLLM(prompt, {
-        userId: 'system',
-        conversationId: 'sentiment-analysis',
-        messageHistory: [],
-      });
-
-      // Parse JSON response or return default
-      try {
-        return JSON.parse(response);
-      } catch {
-        return {
-          overall: 'neutral',
-          confidence: 50,
-          emotions: ['uncertain'],
-          keyPhrases: [text.split(' ').slice(0, 3).join(' ')],
-        };
-      }
-    } catch (error) {
-      this.logger.error('Error in sentiment analysis', { error });
-      throw error;
-    }
-  }
-
-  /**
-   * Tip content creator with USDC
-   */
-  private async tipCreator(
-    creatorAddress: string, 
-    amount: number, 
-    contentId?: string, 
-    message?: string
-  ): Promise<string> {
-    if (!this.agentKit) {
-      throw new Error('AgentKit not initialized');
-    }
-
-    try {
-      const prompt = `Send ${amount} USDC to ${creatorAddress} as a tip for content creation${message ? ` with message: ${message}` : ''}`;
-      
-      const response = await this.processWithLLM(prompt, {
-        userId: 'system',
-        conversationId: 'social-tip',
-        messageHistory: [],
-      });
-
-      // Extract transaction hash from response
-      const txHashMatch = response.match(/0x[a-fA-F0-9]{64}/);
-      return txHashMatch ? txHashMatch[0] : response;
-    } catch (error) {
-      this.logger.error('Tip transfer failed', { error });
-      throw error;
-    }
-  }
-
-  /**
-   * Get token sentiment from real data
-   */
-  private async getTokenSentiment(token: string): Promise<{ score: number; source: string }> {
-    try {
-      // Use CoinGecko for sentiment data
-      const response = await axios.get(`https://api.coingecko.com/api/v3/coins/${token.toLowerCase()}`, {
-        params: { localization: false, tickers: false, market_data: true, community_data: true }
-      });
-
-      const data = response.data;
-      let score = 50; // neutral default
-
-      // Calculate sentiment based on price change and social metrics
-      if (data.market_data) {
-        const priceChange24h = data.market_data.price_change_percentage_24h || 0;
-        const priceChange7d = data.market_data.price_change_percentage_7d || 0;
-        
-        score = 50 + (priceChange24h * 2) + (priceChange7d * 0.5);
-        score = Math.max(0, Math.min(100, score));
-      }
-
-      return {
-        score: Math.round(score),
-        source: 'CoinGecko Market Data'
-      };
-    } catch (error) {
-      this.logger.warn('Error fetching token sentiment', { error, token });
-      return { score: 50, source: 'Default' };
-    }
-  }
-
-  /**
-   * Create content summary using AI
-   */
-  private async createContentSummary(url: string, topic?: string): Promise<{
-    title: string;
-    summary: string;
-    keyPoints: string[];
-  }> {
-    try {
-      // Fetch content
-      const response = await axios.get(url, {
-        timeout: 10000,
-        headers: { 'User-Agent': 'SocialAgent/1.0' }
-      });
-
-      const content = response.data;
-      
-      const prompt = `Summarize this web content${topic ? ` focusing on ${topic}` : ''}:
-${content.substring(0, 2000)}
-
-Return as JSON with:
-- title: string
-- summary: string (max 200 words)
-- keyPoints: array of strings`;
-
-      const aiResponse = await this.processWithLLM(prompt, {
-        userId: 'system',
-        conversationId: 'content-summary',
-        messageHistory: [],
-      });
-
-      try {
-        return JSON.parse(aiResponse);
-      } catch {
-        return {
-          title: 'Content Summary',
-          summary: 'Unable to parse content summary',
-          keyPoints: ['Error processing content']
-        };
-      }
-    } catch (error) {
-      this.logger.error('Error creating content summary', { error, url });
-      throw error;
-    }
-  }
-
-  /**
-   * Track social engagement metrics
-   */
-  private async trackSocialMetrics(contentType: string, timeframe: string): Promise<{
-    topContent: { title: string; engagement: number };
-    totalEngagement: number;
-    viralScore: number;
-    trendingHashtags: string[];
-  }> {
-    try {
-      // This would integrate with real social media APIs in production
-      // For now, providing realistic simulated data based on real patterns
-      
-      const mockMetrics = {
-        topContent: {
-          title: `${contentType === 'defi' ? 'DeFi Protocol Reaches $1B TVL' : 'Bitcoin Breaks Resistance Level'}`,
-          engagement: Math.floor(Math.random() * 10000) + 5000
-        },
-        totalEngagement: Math.floor(Math.random() * 50000) + 25000,
-        viralScore: Math.floor(Math.random() * 40) + 60,
-        trendingHashtags: [
-          '#crypto', '#blockchain', '#defi', '#web3', '#bitcoin'
-        ].slice(0, 3)
-      };
-
-      return mockMetrics;
-    } catch (error) {
-      this.logger.error('Error tracking social metrics', { error });
-      throw error;
-    }
-  }
-
-  protected async handleMessage(message: DecodedMessage, context: AgentContext): Promise<AgentResponse> {
-    try {
-      if (!this.reactAgent) {
-        throw new Error('React agent not initialized');
-      }
-
-      const messageContent = typeof message.content === 'string' ? message.content : '';
-      const threadId = context.conversationId || 'default';
-
-      // Use React Agent to process the message with proper LLM + tool integration
-      let response = '';
-      const stream = await this.reactAgent.stream(
-        { messages: [new HumanMessage(messageContent)] },
-        { configurable: { thread_id: threadId } }
-      );
-
-      for await (const chunk of stream) {
-        if (chunk && typeof chunk === 'object' && 'agent' in chunk) {
-          const agentChunk = chunk as {
-            agent: { messages: Array<{ content: unknown }> };
-          };
-          response += String(agentChunk.agent.messages[0].content) + '\n';
-        }
-      }
-
-      return {
-        message: response.trim() || 'I can help you with crypto news, social content, and community engagement. What would you like to explore?',
-        metadata: { 
-          handledBy: 'social-agent',
-          walletAddress: this.walletProvider ? this.walletProvider.getAddress() : null,
-          usedReactAgent: true
-        },
-        actions: []
-      };
-    } catch (error) {
-      this.logger.error('Error in handleMessage', { error });
-      return {
-        message: `Sorry, I encountered an error: ${error instanceof Error ? error.message : 'Unknown error'}. Please try again.`,
-        metadata: { handledBy: 'social-agent', error: true },
-        actions: []
-      };
-    }
-  }
-
-  protected async shouldHandleMessage(message: DecodedMessage, context: AgentContext): Promise<boolean> {
-    const content = typeof message.content === 'string' ? message.content.toLowerCase() : '';
-    const socialKeywords = [
-      'news', 'trending', 'content', 'social', 'feed', 'recommend', 
-      'crypto news', 'updates', 'sentiment', 'community', 'share',
-      'tip', 'donate', 'support', 'creator', 'viral', 'hashtag'
-    ];
-    
-    return socialKeywords.some(keyword => content.includes(keyword));
-  }
-
-  protected async suggestNextAgent(message: DecodedMessage, context: AgentContext): Promise<string> {
-    const content = typeof message.content === 'string' ? message.content.toLowerCase() : '';
-    
-    if (content.includes('trade') || content.includes('defi') || content.includes('swap')) return 'trading';
-    if (content.includes('game') || content.includes('play')) return 'game';
-    if (content.includes('event') || content.includes('payment') || content.includes('split')) return 'utility';
-    if (content.includes('app') || content.includes('tool') || content.includes('calculator')) return 'miniapp';
-    
-    return 'master';
-  }
-
-  private initializeContentSources(): void {
-    // Real content sources for production use
-    const sources: ContentSource[] = [
-      {
-        name: 'CoinGecko',
-        type: 'api',
-        url: 'https://api.coingecko.com/api/v3/news',
-        categories: ['bitcoin', 'ethereum', 'defi', 'nft'],
-        isActive: true,
-      },
-      {
-        name: 'CryptoCompare',
-        type: 'api',
-        url: 'https://min-api.cryptocompare.com/data/v2/news/',
-        categories: ['altcoins', 'blockchain', 'regulation'],
-        isActive: true,
-      },
-      {
-        name: 'CoinDesk',
-        type: 'rss',
-        url: 'https://www.coindesk.com/arc/outboundfeeds/rss/',
-        categories: ['institutional', 'funding', 'technology'],
-        isActive: true,
-      },
-    ];
-
-    sources.forEach(source => this.contentSources.set(source.name, source));
-  }
-
-  private isNewsRequest(content: string): boolean {
-    return /\b(news|article|update|headline|breaking)\b/.test(content);
-  }
-
-  private isContentRequest(content: string): boolean {
-    return /\b(content|feed|recommend|summary|curation)\b/.test(content);
-  }
-
-  private isTrendingRequest(content: string): boolean {
-    return /\b(trending|viral|popular|hot|buzz)\b/.test(content);
-  }
-
-  private isTippingRequest(content: string): boolean {
-    return /\b(tip|donate|support|pay|reward)\b/.test(content);
-  }
-
-  private async handleNewsRequest(message: DecodedMessage, context: AgentContext): Promise<AgentResponse> {
-    const prompt = `Get the latest crypto news for: "${message.content}". Use real news sources to provide current information.`;
-    const response = await this.processWithLLM(prompt, context);
-    
-    return {
-      message: response,
-      metadata: { handledBy: 'social-agent', category: 'news' },
-      actions: [
-        {
-          type: 'notification',
-          payload: { 
-            message: 'Fetching latest crypto news',
-            source: 'real_apis'
-          }
-        }
-      ]
-    };
-  }
-
-  private async handleContentRequest(message: DecodedMessage, context: AgentContext): Promise<AgentResponse> {
-    const prompt = `Curate content based on: "${message.content}". Use real content sources and provide valuable recommendations.`;
-    const response = await this.processWithLLM(prompt, context);
-    
-    return {
-      message: response,
-      metadata: { handledBy: 'social-agent', category: 'content' },
-      actions: []
-    };
-  }
-
-  private async handleTrendingRequest(message: DecodedMessage, context: AgentContext): Promise<AgentResponse> {
-    const prompt = `Show trending topics for: "${message.content}". Use real social data and market information.`;
-    const response = await this.processWithLLM(prompt, context);
-    
-    return {
-      message: response,
-      metadata: { handledBy: 'social-agent', category: 'trending' },
-      actions: []
-    };
-  }
-
-  private async handleTippingRequest(message: DecodedMessage, context: AgentContext): Promise<AgentResponse> {
-    const prompt = `Handle tipping request: "${message.content}". Use real USDC transfers to support content creators.`;
-    const response = await this.processWithLLM(prompt, context);
-    
-    return {
-      message: response,
-      metadata: { handledBy: 'social-agent', category: 'tipping' },
-      actions: [
-        {
-          type: 'transaction',
-          payload: { 
-            request: message.content,
-            type: 'social_tip'
-          }
-        }
-      ]
-    };
-  }
-
-  protected getSystemPrompt(): string {
-    return `You are SocialAgent, a production-grade social content curator and community engagement specialist powered by real APIs and Coinbase AgentKit.
-
-Your capabilities include:
-- Real-time crypto news aggregation from multiple sources (CoinGecko, CryptoCompare, CoinDesk)
-- Live trending topic analysis from social platforms
-- AI-powered sentiment analysis and content summarization
-- Social tipping with USDC to support content creators
-- Market sentiment tracking for cryptocurrencies
-- Viral content detection and social metrics
-
-You have access to REAL data sources and blockchain tools:
-- Live news APIs for current crypto information
-- Real market data for sentiment analysis
-- USDC tipping functionality via AgentKit
-- Social engagement tracking across platforms
-- AI content summarization and curation
-
-Guidelines:
-1. Always use real, current data from live APIs
-2. Provide accurate news and trending information
-3. Analyze sentiment using AI-powered tools
-4. Support content creators through blockchain tipping
-5. Track and report real social engagement metrics
-6. Curate valuable content based on user interests
-7. Maintain transparency about data sources
-
-Current network: ${process.env.NETWORK_ID || 'base-sepolia'}
-You provide real social intelligence and blockchain-powered community engagement.`;
-  }
-}
